pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = "Voice-first AI orchestrator — macOS menu bar app"
requires-python = ">=3.11"
dependencies = [
    # Menu bar
    "rumps>=0.4.0",
    # Hotkey
    "pynput>=1.7.7",
    # Audio capture (push-to-talk, no VAD needed)
    "sounddevice>=0.4.6",
    "numpy>=1.26",
    "scipy>=1.11",
    # STT
    "deepgram-sdk>=3.0.0",
    # LLM orchestration
    "openai>=1.30.0",
    # Email drafting
    "google-generativeai>=0.7.0",
    "google-api-python-client>=2.130.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    # Web search
    "tavily-python>=0.3.0",
    # TTS
    "elevenlabs>=1.2.0",
    # LangGraph orchestration + conversation memory
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    # PDF text extraction
    "pdfplumber>=0.10.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    # Screenshot (macOS only)
    "pyobjc>=9.0; sys_platform == 'darwin'",
    "Pillow>=10.0.0",
    "deepagents>=0.5.3",
    "langchain-anthropic>=1.4.1",
    "langchain-openai>=1.1.14",
]

[project.scripts]
friday = "friday.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["friday"]
